refactor(library/Attempt): moves "*/Error/modifier" to "*/Error/modifiers"
