+++ src/library/Attempt/Error/exports.object.auxiliaries.ts
@@ -1,1 +1,1 @@
-export * from './modifier';
+export * from './modifiers';
