+++ src/library/Attempt/Error/modifiers/AppropriatelyThrown/assume.ts
