+++ src/library/Attempt/Error/modifiers/AppropriatelyThrown/definition.declared.ts
