+++ src/library/Attempt/Error/modifiers/AppropriatelyThrown/exports.object.primary.ts
