+++ src/library/Attempt/Error/modifiers/AppropriatelyThrown/index.ts
