+++ src/library/Attempt/Error/modifiers/PotentiallyActionable/assume.ts
